Do not try to initialize async fixtures without explicit asyncio mark in strict mode (#307)
* test: Package-scoped event_loop fixture no longer leaks into other tests.

The expected behaviour is that the `event_loop` fixture defined in `tests/sessionloop/conftest.py` is torn down when all tests in `tests/sessionloop` are complete. Running the tests with the pytest option --setup-show pointed out that the fixture is torn down at the end of the test session, instead. This is an unintended side effect of the sessionloop test which may affect other tests in the test suite.

Reducing the fixture scope from "package" to "module" results in the expected behaviour. The module was renamed to reflect the fact that the tests do not use a session scope.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>

* test: Removed test with obsolete "forbid_global_loop".

forbid_global_loop was an option to pytest.mark.asyncio which was removed in v0.6.0. The two subprocess tests are otherwise identical. Therefore, one of the tests was removed along with the obsolete option.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>

* test: Ignore subprocess tests when running on CPython 3.7.

When run with Python 3.7 asyncio.subprocess.create_subprocess_exec seems to be
affected by an issue that prevents correct cleanup. Tests using pytest-trio
will report that signal handling is already performed by another library and
fail. [1] This is possibly a bug in CPython 3.7, so we ignore this test for
that Python version.

CPython 3.7 uses asyncio.streams.StreamReader and asyncio.streams.StreamWriter
to implement asyncio.streams.StreamReaderProtocol and
asyncio.subprocess.SubprocessStreamProtocol. StreamReaderProtocol contained
cyclic references between the reader and the protocol, which prevented
garbage collection. While StreamReaderProtocol received a patch [2],
SubprocessStreamProtocol, which is used by create_subprocess_exec, possibly
has the same problem, but was not patched as part of CPython 3.7.

That's why we ignore this test for CPython 3.7.

[1] https://github.com/python-trio/pytest-trio/issues/126
[2] https://github.com/python/cpython/pull/9201

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>

* build: Added pytest-trio to the test dependencies.

This allows testing compatibility between pytest-trio and pytest-asyncio.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>

* fix: Do not try to initialize async fixtures without explicit asyncio mark in strict mode.

This fixes a bug that breaks compatibility with pytest_trio.

Closes #298

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>