Remove setup.py (#315)
* build: Remove deprecated setup.py.

setup.py is no longer necessary when the project is installed with a PEP 517 compatible build. Pip has support for PEP 517 since v19.0.

The setuptools documentation warns about removing setup.py when editable installations need to be possible. This shouldn't be an issue since editable installations are still possible with pip v21.1 or newer. [1]

I don't think a bump of the minimal setuptools version is necessary for this change.

Closes #283

[1] https://pip.pypa.io/en/stable/news/#v21-1

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>

* docs: Added changelog entry about removal of setup.py.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>