refactor: Extracted logic for marking tests in auto mode into pytest_collection_modifyitems.
pytest_pycollect_makeitem currently calls `Collector._genfunctions`, in order to delegate further collection of test items to the current pytest collector. It does so only to add the asyncio mark to async tests after the items have been collected.

Rather than relying on a call to the protected `Collector._genfunctions` method the marking logic was moved to the pytest_collection_modifyitems hook, which is called at the end of the collection phase. This change removes the call to protected functions and makes the code easier to understand.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>