feat!: Remove legacy mode.
Asyncio modes have been introduced in v0.17.0 which was released on 2022-01-13. In addition to the auto and strict modes, the release defaulted to "legacy" mode for backwards compatibility. The release also issued a deprecation warning when using legacy mode.

The default asyncio mode was changed to _strict_ in the v0.19.0 release on 2022-07-13.

This change removes legacy mode entirely.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>