[ci] Report combined coverage across all interpreter versions.
Some codein pytest-asyncio is specific to certain versions of the Python interpreter. Therefore, the code coverage can be different for tests runs with different interpreters. Currently, the CI pipeline does not accommodate this fact.

This commit changes coverage data to be generated by each CI environment during the test runs. The different coverage results are combined in the check step just before uploading them to codecov.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>