[test] Remove pytest-trio from test dependencies.
trio v0.22 deprecated trio.MultiError. This adjustment requires at least pytest-trio v0.8.0, [0] which, in turn, depends on pytest>=7.2. Pytest-asyncio currently supports pytest>=7.0. Therefore, we either need to juggle with the constraints in the pytest-min test environment or remove the pytest-trio dependency entirely.

I opted for the latter. Pytest-asyncio should be compatible with other plugins, but it should not be responsible to test that compatibility. If every pytest plugin wrote tests against other plugins, we ended up in dependency hell and a lot of work would be done multiple times.

There's a dedicated section in pytest for testing plugin integration. This is the right place to do it.

[0] https://pytest-trio.readthedocs.io/en/v0.8.0/history.html#misc

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>