[fix] Fixes a bug related to the ordering of the "event_loop" fixture in connection with parametrized tests.
The fixture evaluation order changed for parametrizations of the same test. The reason is probably the fact that `event_loop` was inserted at position 0 in the pytest fixture closure for the current test. Since the synchronization wrapper for async tests uses the currently installed event loop rather than an explicit reference as of commit 36b226936e17232535e88ca34f9707cdf211776b, we can drop the insertion of the event_loop fixture as the first fixture to be evaluated.

This patch also addresses an issue that caused RuntimeErrors when the event loop was set to None in a fixture that is requested by an async test. This can occur due to the use of asyncio.run, for example.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>