[fix] Fixes a bug that caused internal pytest errors during test collection with doctests.
Pytest-asyncio attaches an event loop of a particular scope to each collector during pytest_collectstart. The implementation did not account for the fact that collectors may have specialized subclasses, such as DoctestModule. This caused a KeyError during a dictionary access, leading to an internal pytest error in the collection phase.

This patch changes the implementation of pytest_collectstart to account for subclasses of collectors.

Signed-off-by: Michael Seifert <m.seifert@digitalernachschub.de>